+++ src/CoordinateOperation/Params/EPSG10515.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * PHPCoord.
+ *
+ * @author Doug Wright
+ */
+declare(strict_types=1);
+use PHPCoord\UnitOfMeasure\UnitOfMeasureFactory;
+
+/** @internal */ return [
+  'latitudeOfNaturalOrigin' => UnitOfMeasureFactory::makeUnit(44.151200716, 'urn:ogc:def:uom:EPSG::9110'),
+  'longitudeOfNaturalOrigin' => UnitOfMeasureFactory::makeUnit(-90.503946745, 'urn:ogc:def:uom:EPSG::9110'),
+  'scaleFactorAtNaturalOrigin' => UnitOfMeasureFactory::makeUnit(1.0000353, 'urn:ogc:def:uom:EPSG::9201'),
+  'falseEasting' => UnitOfMeasureFactory::makeUnit(88582.5, 'urn:ogc:def:uom:EPSG::9003'),
+  'falseNorthing' => UnitOfMeasureFactory::makeUnit(82020.833, 'urn:ogc:def:uom:EPSG::9003'),
+];